Add timestamps to trace-db-queries.d (#7405)
It can be difficult to correlate the raw queries captured by the
`trace-db-queries.d` script with external events without knowing the
time the query completed.

Add microsecond-precision timestamps at the start of each event. To keep
the script simple, do this by appending fractional seconds to the standard
`%Y` format.

    2025 Jan 27 22:35:26.531703 - conn_id: a2c53734-...