[reconfigurator] Move insertions into `crucible_dataset` from executor RPW to rendezvous RPW (#7397)
This PR builds on #7386. I'll test both of them together on a racklette
(will put results in a comment on this PR).

This PR contributes to addressing (but doesn't fully address) a couple
issues:

* #7392: This PR adds useful output for the `crucible_dataset`
rendezvous table, but we still need useful output for `debug_dataset`
* #6999: This PR removes a step from execution that currently depends on
successful completion of earlier step to post datasets to sled-agents,
since we now confirm insertions are okay via inventory collection

I'm inclined to say this PR fully addresses a couple other issues, but
it's debatable:

* #7210: Are we okay with IP and port still being here now that this
table is just crucible datasets? (IMO it's still conceptually weird to
say a dataset has an IP/port, but maybe it's not as bad now?)
* #6998 `dataset` as a general table is now gone and replaced with more
specific rendezvous tables, so I think this one can be closed?