nexus: split out control plane zones as separate artifacts (#7452)
The primary change here is adding a control knob to
`ArtifactsWithPlan::from_stream` that either splits out the artifacts or
combines all the zones into a single tarball.

`ControlPlaneZonesMode::Composite` is the current behavior and is used
in Wicket and existing unit tests. For the moment this doesn't support
handling repositories with artifacts already split out, but the
semantics of this setting are that if/when we do that, it will combine
them into a single artifact for Wicket.

`ControlPlaneZonesMode::Split` enables the new behavior and is used in
Nexus. This makes the individual zone images available in the TUF Repo
Depot and therefore immediately usable by Sled Agent when creating
zones.

Additionally, we add a check to assembling a TUF repo to ensure that no
individual zone images have the same hash; since the fake zones created
during tests had the same hashes, we learned such a repo would not be
accepted by Nexus due to this check:


https://github.com/oxidecomputer/omicron/blob/e5bf48fdba89707257ef6ba75375714359941370/update-common/src/artifacts/update_plan.rs#L875-L882

Closes #4411. (If we want to split out the control plane zones in actual
repos we can open a new issue for that.)