[reconfigurator] Move resource allocation out of `BlueprintBuilder` (#7235)
Most of this diff is moving stuff around in preparation for a larger
rework, but there are some nontrivial changes worth reviewing that I'll
call out in comments below. This is staged on top of #7234.

Prior to this PR, `BlueprintBuilder` had 3 fields that allocated various
kinds of resources:

* `sled_ip_allocators` (for allocating underlay IPs)
* `external_networking` (for creating OPTE NICs and assigning external
IP addresses for Nexus/boundary NTP/external DNS)
* `internal_dns_subnets` (for assigning internal DNS zones, which get
IPs distinct from the normal sled subnet)

All three were lazily instantiated on first use. This is to allow the
planner to first expunge zones then start adding; the first "add" that
needs a resource from one of these allocators creates the allocator by
snapshotting the state of available resources at that point (_after_
expungement). Eventually I'd like to break the builder up to make these
two phases more explicit, but for now this PR settles for shrinking the
builder's surface area some:

* `sled_ip_allocators` is gone; each `SledEditor` now maintains its own
underlay IP allocator. These are no longer instantiated lazily, but for
underlay IPs specifically this is fine: we never reuse underlay IPs from
expunged zones, so we don't need to wait for expungement to reuse
resources like the other two.
* `external_networking`/`internal_dns_subnets` were replaced by
`resource_allocator`, which internally holds both of those types and has
pass-through methods for the builder to use. This is still lazily
instantiated via `OnceCell` to keep the existing implicit
ordering/behavior.