[oximeter] Fix self_stat count test flakes (#7456)
#7262 tried to fix the off-by-one test flake #7255, but it's still
around. I think based on discussion on that PR
(https://github.com/oxidecomputer/omicron/pull/7262/files#r1890548730),
the assertion in it was adding one to the wrong side: it's possible the
fake server may have seen one more request than the agent has finished
noting, which means `server_count` may be equal to either `count` or
`count + 1`, not `count` or `count - 1`.

The change from `assert_eq` to `assert` also caused the new flakes to
not log the actual values; this puts them back into the assertion
message in case this fix is still not quite right.