blueprint-builder: Don't drop decommissioned sleds out of `sled_state` (#7470)
This is one part working towards #7078 and #7309 (and a few related
issues). We'd like to be able to send sled-agent a complete config
containing its zones+datasets+disks, which means first we need to merge
the separate maps in the blueprint together (and attach an overriding
generation number on this combo sled config).

Before we can do _that_, we need to ensure that each of the four maps
keyed by a sled ID in the blueprint (`sled_state`, `blueprint_zones`,
`blueprint_disks`, `blueprint_datasets`) have matching sets of keys.
Viewing `blueprint_zones` as the primary map,
`BlueprintBuilder::build()` has three "let's match the existing
behavior" spots that can _drop_ entries from the other three maps.

This PR removes one of those three, and ensures that any sled ID present
in `blueprint_zones` is also present in `sled_state`. (Previously, we'd
drop sleds from `sled_state` after decommissioning.)

Most of the PR bulk is adding new checks to blippy to look for
inconsistencies between the four maps.

---------

Co-authored-by: Andrew J. Stone <andrew@oxidecomputer.com>