[reconfigurator] blueprint builder: backfill filesystem_pool (#7450)
This builds on #7449 and, in combination with it, fixes #7229.

When planning a new blueprint, one of the first things we now do is loop
over all in-service zones on all active sleds and check that their
`filesystem_pool` is consistent with the latest inventory collection
from that sled. The expectation here is that we'll be filling in
`filesystem_pool: None` values from older systems (i.e., #7229). There's
also an edge case where we might need to correct our own incorrect
backfills, if a sled-agent has restarted since the inventory collection
was taken and has chosen a different zpool for some zone.

If we can't find what the correct filesystem pool should be (e.g., if
the sled isn't present in inventory at all, or none of its zpool/dataset
combinations match the zone), we leave the `filesystem_pool` value
untouched.

I'll put testing notes from a4x2 in a comment below.