[reconfigurator] Add target blueprint to reconfigurator state (#7485)
This PR adds the target blueprint as an optional field in
`UnstableReconfiguratorState`. I don't love that it's optional, because
on any real deployed system, there is guaranteed to be a target
blueprint. However, `UnstableReconfiguratorState` is also used by
`reconfigurator-cli`, and it has various ways for a system state in
memory to have no target blueprint. E.g., it has a test where it starts
from an empty state, adds a sled, then writes out an
`UnstableReconfiguratorState` that contains that sled but no blueprints
at all (and therefore no target blueprint).

I'm not sure how much fidelity reconfigurator-cli's simulated system
should be required to follow, but it seemed like a nontrivial lift to
require it to always have a blueprint, hence making the field optional.
Maybe it could use a different serialization struct, but that seems not
ideal in other ways. Very open to ideas!

Builds on #7484. PR 2 of 3 moving toward adding `omdb db reconfigurator
archive`.