[omdb] Add datasets to `blueprints show ...` (#7506)
A couple times debugging this week I tried to `omdb nexus blueprints
show ...` to look at datasets, and had to resort to looking at diffs or
raw JSON, so I figured it was time to flesh this out. The diff is almost
all expectorate output; code changes are pretty small. I tried to make
it slightly less likely we forget to update the blueprint display when
fields are changed in the future.

Fixes #7303.