Region snapshot replacement for read-only regions (#7435)
Reuse the region snapshot replacement machinery to replace read-only
regions. This is done by storing a replacement type in the region
snapshot replacement record such that either a region snapshot _or_ a
read-only region can be the subject of this type of replacement. The
procedure for both types is the same so all the code can be reused.

A future commit will rename region snapshot replacement (and all
references) to "read-only target replacement" to reflect that the
machinery now applies to both region snapshots and read-only regions.
This will be a mostly mechanical set of changes that can be reviewed
separately with much less scrutiny. Right now manually requesting a
region replacement with omdb is done through the region replacement
manual request, not the region snapshot replacement manual request. This
will change in that future commit to be part of a read-only target
replacement request.

Fixes #6172