Routing RPW: Allow only `ObjectNotFound` for lookup failures
This stems from part of #7534. In understanding that, we have a
hypothetical failure mode where any lookup which fails due to a database
failure (rather than a name lookup) could lead to gaps in routes, which
would require a router bump to retry.

This PR changes resolution of all items such that any unexpected DB failures
(transient or otherwise) *will* cause the router's resolution to bail. It
will take some time before it's recomputed, but we know it will be done.