Address `test_region_replacement_triple_sanity_2` flake
`test_region_replacement_triple_sanity_2` performs two disk expungements
before starting the Crucible replacement machinery to ensure that losing
two out of three parts of a Volume's region set will not result in lost
data. This test was occasionally failing with an assertion that
indicated not all resources on expunged physical disks had been replaced
in a volume.

There are two fixes in this commit:

The original guess as to why this test was failing was that part of the
`wait_for_all_replacements` loop that checked how many "actions" were
taken as when activating all the related background tasks with
`run_all_crucible_replacement_tasks` was racing with another (probably
periodic) activation of those background tasks, and therefore returning
a previous activation's status that said no actions were taken. This
commit changes `activate_background_task` so that it will only exit when
it has activated the requested background task.

This is not the cause of the flake however: `wait_for_all_replacements`
is a routine that previously would:

1. run all background tasks related to region replacement and region
   snapshot replacement

2. bail out of the loop if there are no more non-completed replacements

This is incorrect! "no more non-completed replacements" is not an
indicator that all necessary resource movement has completed, only that
_current_ movement has completed. Rerunning the background tasks could
request more work.

The real check that function should be performing is whether or not any
resources used by Volumes are backed by physical disks that are
expunged. This commit changes `wait_for_all_replacements` to:

1. run all background tasks related to region replacement and region
   snapshot replacement, but ignore the output

2. check how many resources are still backed by expunged physical disks,
   and loop if there still are some

3. check if work is in-progress, and loop if so

4. return to the caller

Also in this commit:

- an omdb command has been added that will print what resources need to
  be replaced (because they're on expunged physical disks)

- an errant eprintln has been removed (it was showing up in the
  `--nocapture` output)

Fixes #7535