prioritize some slow tests to run first (#7542)
I've noticed some tests be really slow, and if we decide to run them at
the end
can cause the entire suite to take longer. Use nextest's new
prioritization
feature to run them first.

A CI test run:


[Before](https://buildomat.eng.oxide.computer/wg/0/details/01JM0MBHDGSK6XN3HBEHCKZGVD/jAmTO9CSgt66QfrBETI1Iu6LV1AmFAagcLSucIFVsgQnRmLG/01JM0MBWCAGBTDYV95YBBH6FHD#S7624):

> Summary [2608.012s] 1802 tests run: 1802 passed (23 slow, 1 leaky), 4
skipped


[After](https://buildomat.eng.oxide.computer/wg/0/details/01JM1FZ1EJW8WC53MNAXSTQFG4/L4NZVElgTNha3FQe1HR2B9AKEdlEmOBNvbbuuu4BVxtHsTZz/01JM1FZFRZEZZ8PTN4HHYDJSKX#S7626):

> Summary [2182.362s] 1802 tests run: 1802 passed (21 slow, 1 leaky), 4
skipped

That's an around 400 second speedup -- not bad! There's still some
opportunity in parallelizing the ClickHouse cluster tests, but this is
good for now.