[reconfigurator] Fix assert_planning_makes_no_changes() (#7512)
Despite its name, this test function didn't actually invoke the planner.
Now it does, and this fixes up some tests that now fail (and changes the
behavior of the planner in one case where the test failure arguably
indicated some planner misbehavior).

The specific planner change: Previously, if we added an NTP zone to a
sled, we considered that sled ineligible for any other zones (consistent
with the sled-agent behavior of refusing to start zones if time isn't
sync'd yet). This PR relaxes that slightly: if we're adding an NTP zone
to a sled and that sled already has other zones that needed time to be
sync'd, we do consider it eligible. This avoids having to make multiple
planner iterations to repopulate a sled that had a subset of disks/zones
expunged, while maintaining the existing behavior for new sleds being
added.

I'm not totally convinced that the existing behavior for new sleds is
necessary; since we're only waiting for the presence of an NTP zone, not
an inventory report that time is actually sync'd, I'm not sure there's a
meaningful difference to requiring two planning iterations instead of
one to place extra zones. But I don't have a reason to propose changing
it or testing it other than "seems a little over-cautious", so it seems
fine to leave it.