Show more information in omdb db snapshots (#7504)
Now we print, when available, the region snapshot info as well as the
datasets that make up the source.
Changed the output a little to fit in fewer columns.

Having this info available in omdb allows us to write tests to find an
replace a specific part of a downstairs in a snapshot.

---------

Co-authored-by: Alan Hanson <alan@oxide.computer>