Support multicast with `sp-sim` (#7545)
Previously the simulated SP did not truly support listening in a
multicast group. This commit changes the network configuration for the
simulated SP so that either "simulated" or "real" is explicitly
configured:

- if "simulated", then most of the current behaviour is preserved:
listen on a bind address but don't join a multicast group. Note: the
current behaviour when `multicast_addr` is supplied is incorrect because
without the interface index matching the interface that is physically
connected to MGS (or faux-mgs), it doesn't work.

- if "real", then require all the necessary parameters to correctly
configure joining a multicast group for an interface, and set it up.