Simplify system internet gateway targets (#7553)
This PR is best viewed as two commits. The first does a bit of cleanup
around platform-specific code to make it easier to test our OPTE
`PortManager`, and then introduces a regression test for #7541. That
test sets up the specific sequence of events we believe triggers the
issue, which is due to incoherent state between the routes the
`PortManager` believes it has and those it's actually applied in OPTE.
This commit also includes commented-out code that _passes_ -- that's
useful for showing the current (incorrect) behavior, and for writing the
test itself.

The second commit fixes #7541, by simplifying how we refer to routes
that target an Internet Gateway. Rather than allow an optional ID of the
gateway, we make explicit that we're targeting either the builtin
services VPC's internet gateway or a customer / instance VPC's gateway.
In the latter, we always include the ID of the gateway, which is how
it's currently used on `main` anyway.

In the former, we _never_ include the ID. During bootstrapping of the
system (e.g., NTP sync), we don't know the gateway's ID yet, because
it's part of the "fixed database data" that we only learn after starting
up Nexus. And even when we do learn the ID, it's a static, fixed ID and
there's only one of them, so it's not particularly useful. But by
removing the ID, we ensure that our logic in the `PortManager` for
deciding what new routes to apply or remove on a notification from
Nexus, we don't confuse ourselves by making the (useless) distinction
between a route targeting the services VPC Internet Gateway just because
one request has an ID and the other does not.