Expunge and Decommission disks in planner (#7286)
This PR partially fixes #6999,  and fixes #7098 and #7502.

A few significant changes are made in this PR:

 1.  Where possible execution steps no longer error out and cause
the executor itself to early return without running all steps. Instead a
`StepWarning` is returned for each step that contains errors and the
executor
runs every step.
2. Some fallible steps have become decoupled in that they do not rely on
each other and can safely run independently. This was mostly true,
except for
the bulk of the work in this PR which ensures that when a a disk is
expunged, it
is not immediately decommissioned unless the sled is also expunged. The
planner
waits to see that the expungement has been observed by a sled-agent on
an active
sled with an expunged disk before it marks it decommissioned.
3. The executor only decommissions physical disks if their disposition
is expunged and they are `ready_for_cleanup` in the blueprint. This
completes the feedback loop
between planning, inventory collection, and execution. Previously disk
decommissioning was done for *all* expunged disks. This was racy in
that a separate nexus may have expunged a disk that was not yet ready
to be decommissioned, and it would be picked up in the decommissioning
step of the first nexus's executor. There is more information in the
following [commit
comments](https://github.com/oxidecomputer/omicron/commit/3ee1546af757e9f7c396794057ca83cac68fac67#commitcomment-149432816).
4. A Disposition field have been added to the table output for disks.

Steps 1 and 2 above partially resolve #6999 and resolve #7502. While I
was modifying the datastore
to implement step 3 I also updated the `modified_time` of the physical
disks
table for necessary transactions. This fixes #7502.

This PR also fixes a planner bug where we only check the planning input
for
decommissioned disks even if there was a change in the parent blueprint,
but no
corresponding change to the planning input.