[reconfigurator] Planner: mark zones `ready_for_cleanup` when appropriate (#7588)
This PR implements two related planner changes:

* When expunging a sled, all the zones are immediately marked as
`Expunged { ready_for_cleanup: true, .. }`.
* When there are expunged zones on an in-service sled, we'll consult the
latest inventory collection and flip `ready_to_cleanup` to true if the
zone is gone and the sled's zone config generation is >= the generation
in which the zone was expunged.

No downstream users of `ready_for_cleanup` are included in this PR;
those will come in followups.