[reconfigurator] Don't reuse internal DNS subnets until previous zone is ready for cleanup (#7589)
Fixes #7380.

Builds on #7588. In addition to the primary change, includes a couple
smaller things worth a second look:

* Lifts the check on the policy for the number of internal DNS zones out
of `InternalDnsSubnetAllocator::new()` (which was kinda dubious, since
the allocator doesn't actually care about the policy) and into a new
method on the planner, called at the top of planning. I _think_ this
would be a reasonable place to put blippy checks, too, if we want to
reject planning on blueprints that contain fatal errors?
* Adds a `BlueprintZoneDisposition::could_be_running()` filter method
for "in service or expunged but not yet ready for cleanup", which is
exactly the check we need for "is it safe to reuse an internal DNS
subnet". I know some folks don't love "could" or "might"; happy to take
naming suggestions here.