[reconfigurator] Planner: Don't omit expunged sleds from `blueprint_datasets` (#7613)
This removes the last "for backwards compatibility..." bit from
`BlueprintBuilder::build()` where we potentially drop entries from one
of the four blueprint maps. The code changes are small:

* Remove the backcompat code
* Fixup some tests
* Address some todos in blippy related to this backcompat implementation
(blippy was only checking maps conditional on other maps; now it always
checks all of them)

and the bulk of the diffs are expectorate changes.