Validate transit IPs on network interface update (#7559)
This PR adds in some checks when setting transit IPs on a NIC to ensure
that we have:
* Only unicast addresses.
* No loopback addresses.
* No duplicates.

As discussed in the ticket, violation of any of these isn't going to
leave OPTE or CRDB in a broken state -- just a confusing one for end
users. This should make things unambiguous in terms of  how the
transit IPs will be handled.

Closes #7530.