[reconfigurator] add image_source to BlueprintZoneConfig and diffs (#7578)
Always set it to `InstallDataset` while reading from the DB for now --
don't store anything in the DB.

Also add some basic support to the blueprint builder for displaying the
image source. I'm a bit concerned about the length of the artifact hash
bloating up the column, but it's something we can deal with in the
future if necessary (maybe by using a short hash and having a
`--verbose` flag for the full hash).